backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.31
python-dotenv==1.0.1
passlib==1.7.4
python-jose==3.3.0
pydantic==2.7.4
arxiv==2.1.0
pypdf==4.3.0
psycopg2-binary==2.9.9
passlib[bcrypt]
pydantic[email]

